Create nix-buildable aoclib package
It's a bit of a WIP right now since tests don't run as is (no hspec dependency)
and when I tried to set up hspec-discover, `cabal test` stopped working.

But it's enough progress, and it builds and runs tests and creates the Haskell library in
the `result/` directory, that I think it's a good stopping place.